Moved event.js to constants folder
