+++ viteapp/src/pages/CalenderPage.jsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import { Calendar, momentLocalizer, Views } from "react-big-calendar";
 import moment from "moment";
 import "react-big-calendar/lib/css/react-big-calendar.css";
-import { MOCK_EVENTS } from "./event";
+import { MOCK_EVENTS } from "../constants/event";
 import { NavBar } from "../components";
 
 const localizer = momentLocalizer(moment);
